Added updated ReadDataFiles file and CreateGameObjects file
The ReadDataFiles.py will read in all the room filed in the rooms folder and return a list of the rooms. It also had function to read in json files  of features and objects, and return a list for them too but I may have to modify it to read from a folder like the rooms one does if we are doing separate files for each item and feature.
CreateGameObjects uses the lists created by ReadDataFiles to create objects and return a list of those objects.